# editor-api/archive_editor/specs/openapi_3_1_0_generated.yaml
openapi: 3.1.0
info:
  title: FastAPI
  version: 0.1.0
paths:
  /edit_simulation/:
    post:
      summary: Edit Archive Simulation
      operationId: edit-simulation
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_edit-simulation'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimulationEditConfirmation'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /download/{file_identifier}:
    get:
      summary: Download Edited File
      description: Serve the edited COMBINE archive for download.
      operationId: download_edited_file_download__file_identifier__get
      parameters:
      - name: file_identifier
        in: path
        required: true
        schema:
          type: string
          title: File Identifier
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    Body_edit-simulation:
      properties:
        archive:
          type: string
          format: binary
          title: Archive
        edit_request_str:
          type: string
          title: Edit Request Str
      type: object
      required:
      - archive
      - edit_request_str
      title: Body_edit-simulation
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    SimulationEditConfirmation:
      properties:
        message:
          type: string
          title: Message
          default: Simulation edited successfully.
        download_link:
          type: string
          title: Download Link
      type: object
      required:
      - download_link
      title: SimulationEditConfirmation
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError